task:  auth

+++ src/main/java/com/shopmanagment/repository/ProductRepositoryImpl.java
@@ -19,26 +19,26 @@ public class ProductRepositoryImpl implements ProductRepository<Product> {
     protected JdbcOperations jdbcOperations;
 
     @Override
-    public void create(Product object) {
-        jdbcOperations.update("INSERT INTO product (shop_id, name, description) VALUES (?, ?, ?);",
-                object.getShop_id(), object.getName(), object.getDescription());
+    public void create(long user_id, Product object) {
+        jdbcOperations.update("INSERT INTO product (user_id, shop_id, name, description) VALUES (?, ?, ?, ?);",
+                user_id, object.getShop_id(), object.getName(), object.getDescription());
     }
 
     @Override
-    public void update(Product object) {
-        jdbcOperations.update("UPDATE product SET name = ?, description = ? WHERE shop_id = ? AND id = ?;",
-                object.getName(), object.getDescription(), object.getShop_id(), object.getId());
+    public void update(long user_id, Product object) {
+        jdbcOperations.update("UPDATE product SET name = ?, description = ? WHERE user_id = ? AND shop_id = ? AND id = ?;",
+                object.getName(), object.getDescription(), user_id, object.getShop_id(), object.getId());
     }
 
     @Override
-    public void delete(long shop_id, long id) {
-        jdbcOperations.update("DELETE FROM product WHERE shop_id = ? AND id = ?;", shop_id, id);
+    public void delete(long user_id, long shop_id, long id) {
+        jdbcOperations.update("DELETE FROM product WHERE user_id = ? AND shop_id = ? AND id = ?;", user_id, shop_id, id);
     }
 
     @Override
-    public List<Product> getProducts(long shop_id) {
+    public List<Product> getProducts(long user_id, long shop_id) {
         List<Product> result = new ArrayList<>();
-        SqlRowSet rowSet = jdbcOperations.queryForRowSet("SELECT * FROM product WHERE shop_id =? ORDER BY id;", shop_id);
+        SqlRowSet rowSet = jdbcOperations.queryForRowSet("SELECT * FROM product WHERE user_id = ? AND shop_id = ? ORDER BY id;", user_id, shop_id);
         while (rowSet.next()) {
             result.add(new Product(
                     rowSet.getLong("id"),
@@ -50,9 +50,9 @@ public class ProductRepositoryImpl implements ProductRepository<Product> {
     }
 
     @Override
-    public Product getProduct(long shop_id, long id) {
+    public Product getProduct(long user_id, long shop_id, long id) {
         List<Product> result = new ArrayList<>();
-        SqlRowSet rowSet = jdbcOperations.queryForRowSet("SELECT * FROM product WHERE shop_id = ? AND id = ?;", shop_id, id);
+        SqlRowSet rowSet = jdbcOperations.queryForRowSet("SELECT * FROM product WHERE user_id = ? AND shop_id = ? AND id = ?;", user_id, shop_id, id);
         while (rowSet.next()) {
             result.add(new Product(
                     rowSet.getLong("id"),
